Fix multiple critical bugs and incomplete features
Backend:
- Fix .decode() crash on already-decoded string in monitoring/backup/update services
- Fix DashboardSummary schema mismatch causing 500 on /dashboard/summary
- Fix FirewallQuickStatus required-vs-None mismatch (hostname, online)
- Fix get_firewall / get_firewall_status returning __dict__ (binary api_secret + SA state)
- Fix /backups restore expecting query param instead of JSON body
- Implement /updates check (was a stub) to actually call OPNsense
- Run backup/update background tasks with own DB session (avoid closed-session crash)
- Fix OPNsenseAPI verify_ssl=True without cert (was silently False)
- Fix dashboard_summary using all status rows instead of latest per firewall
- Use BlockingScheduler instead of BackgroundScheduler + busy-wait
- Make BackupCreate body optional; alerts.list supports resolved=null
- BigInteger for backup size_bytes and uptime_seconds

DB:
- Rewrite schema.sql with valid PostgreSQL (was MySQL INDEX-in-CREATE-TABLE syntax)
- Add missing license_type column

Frontend:
- Fix EMPTY_FORM missing auto_update_window
- Fix nested map() in auto-update window dropdown (use flatMap)
- Parse backup_retention as number not string
- Add Backups link to mobile navbar

+++ backend/app/models/__init__.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, String, DateTime, Boolean, Integer, Float, JSON, Text, ForeignKey, TIMESTAMP, LargeBinary
+from sqlalchemy import Column, String, DateTime, Boolean, Integer, BigInteger, Float, JSON, Text, ForeignKey, TIMESTAMP, LargeBinary
 from sqlalchemy.dialects.postgresql import UUID
 from app.database import Base
 
@@ -57,7 +57,7 @@ class FirewallStatus(Base):
     updates_available = Column(Integer, default=0)
     cpu_usage = Column(Float)
     ram_usage = Column(Float)
-    uptime_seconds = Column(Integer)
+    uptime_seconds = Column(BigInteger)
     gateway_status = Column(JSON)
     pending_services = Column(JSON, default=list)
     last_error = Column(Text, nullable=True)
@@ -71,7 +71,7 @@ class Backup(Base):
     firewall_id = Column(UUID(as_uuid=True), ForeignKey("firewalls.id"), nullable=False, index=True)
     filename = Column(String(255), nullable=False)
     filepath = Column(String(500), nullable=False)
-    size_bytes = Column(Integer)
+    size_bytes = Column(BigInteger)
     created_at = Column(DateTime, default=datetime.utcnow)
     triggered_by = Column(String(50))  # "manual", "auto", "pre-update"
     last_error = Column(Text, nullable=True)
